Refactoring + reorganizing on WeatherService
