+++ MirrorSUPINFO/Sources/Libraries/MirrorSUPINFO.Apps/Modules/Models/WeatherModel/Cloud.cs
@@ -1,0 +1,14 @@
+﻿using Newtonsoft.Json;
+
+namespace MirrorSUPINFO.Apps.Modules.Models.WeatherModel
+{
+    public class Cloud
+    {
+        #region Properties
+
+        [JsonProperty("all")]
+        public int All { get; set; }
+
+        #endregion        
+    }
+}
