+++ MirrorSUPINFO/Sources/Libraries/MirrorSUPINFO.Apps/Modules/Models/WeatherModel/Coord.cs
@@ -1,0 +1,17 @@
+﻿using Newtonsoft.Json;
+
+namespace MirrorSUPINFO.Apps.Modules.Models.WeatherModel
+{
+    public class Coord
+    {
+        #region Properties
+
+        [JsonProperty("lon")]
+        public double Longitude { get; set; }
+
+        [JsonProperty("lat")]
+        public double Latitude { get; set; }
+
+        #endregion       
+    }
+}
