+++ MirrorSUPINFO/Sources/Libraries/MirrorSUPINFO.Apps/Modules/Models/WeatherModel/Weather.cs
@@ -1,0 +1,23 @@
+﻿using Newtonsoft.Json;
+
+namespace MirrorSUPINFO.Apps.Modules.Models.WeatherModel
+{
+    public class Weather
+    {
+        #region Properties
+
+        [JsonProperty("id")]
+        public int Id { get; set; }
+
+        [JsonProperty("main")]
+        public string Main { get; set; }
+
+        [JsonProperty("description")]
+        public string Description { get; set; }
+
+        [JsonProperty("icon")]
+        public string Icon { get; set; }
+
+        #endregion       
+    }
+}
