+++ MirrorSUPINFO/Sources/Libraries/MirrorSUPINFO.Apps/Modules/Models/WeatherModel/Wind.cs
@@ -1,0 +1,17 @@
+﻿using Newtonsoft.Json;
+
+namespace MirrorSUPINFO.Apps.Modules.Models.WeatherModel
+{
+    public class Wind
+    {
+        #region Properties
+
+        [JsonProperty("speed")]
+        public double Speed { get; set; }
+
+        [JsonProperty("deg")]
+        public double Deg { get; set; }
+
+        #endregion       
+    }
+}
