+++ MirrorSUPINFO/Sources/Libraries/MirrorSUPINFO.Apps/project.json
@@ -1,7 +1,8 @@
 ﻿{
   "dependencies": {
     "Microsoft.NETCore.UniversalWindowsPlatform": "5.1.0",
-    "MvvmLightLibs": "5.2.0"
+    "MvvmLightLibs": "5.2.0",
+    "Newtonsoft.Json": "8.0.3"
   },
   "frameworks": {
     "uap10.0": {}
