+++ MirrorSUPINFO/Sources/Libraries/MirrorSUPINFO.Components/ComponentModel/Services/Module/ModuleManager.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using System.Threading.Tasks;
 using Windows.Security.Credentials;
 using Windows.UI.Core;
